TESTS: Fixed main() return type to be int

+++ tests/sail-common/variant.c
@@ -250,6 +250,6 @@ static const MunitSuite test_suite = {
     MUNIT_SUITE_OPTION_NONE
 };
 
-long main(int argc, char *argv[MUNIT_ARRAY_PARAM(argc + 1)]) {
+int main(int argc, char *argv[MUNIT_ARRAY_PARAM(argc + 1)]) {
     return munit_suite_main(&test_suite, NULL, argc, argv);
 }
